more checks in integer validation

+++ src/main/kotlin/io/github/xstefanox/json/schema/validator/model/Multiplier.kt
@@ -20,7 +20,7 @@ data class Multiplier(private val value: Number) : Comparable<Multiplier> {
         return value
     }
 
-    fun divides(num: Int): Boolean {
+    fun divides(num: Long): Boolean {
         return num.rem(value.toDouble()) == 0.0
     }
